Gameplay Balance, enemy dead effect

+++ CGP_Assignment/Code/Object/Tower.h
@@ -43,7 +43,7 @@ public:
 	float distanceY;
 	float distance;
 	bool isFire;
-	int fireCD;
+	float fireCD;
 };
 
 #endif // !TOWER
